+++ ikarus/finiteelements/ferequirements.hh
@@ -204,7 +204,7 @@ auto scalarAffordance(VectorAffordance affordanceV) {
 namespace AffordanceCollections {
   inline constexpr AffordanceCollection elastoStatics(ScalarAffordance::mechanicalPotentialEnergy,
                                                       VectorAffordance::forces, MatrixAffordance::stiffness);
-}
+} // namespace AffordanceCollections
 
 /**
  * \class FERequirements