+++ CHANGELOG.md
@@ -84,6 +84,14 @@ SPDX-License-Identifier: LGPL-3.0-or-later
   ([#337](https://github.com/ikarus-project/ikarus/pull/337))
 - `Observers` and `Observables` are replaced with `Broadcasters` and `Listeners`. Existing loggers work almost the same.
  A noteworthy difference is that, the Broadcaster (e.g. a nonlinear solver) has to be registered to a Listener (e.g. a logger) with `logger.subscribeTo(solver)` ([#349](https://github.com/ikarus-project/ikarus/pull/349))
+- Refactor EAS to handle NonLinearElastic ([#325](https://github.com/ikarus-project/ikarus/pull/325))
+    - This also includes a new interface method in the `Mixin` called `updateState` with the respective `impl()` function.
+    - For EAS, `updateStateImpl()` is used to update the internal variable `alpha_` in a nonlinear analysis.
+    - Missing functions like `getStress` and `materialTangentFunction` are added to `LinearElastic` and `NonLinearElastic`, respectively.
+    - `EnhancedAssumedStrains` class now takes a struct denoting the enhanced strain type as a template argument, which implements the
+      enhanced strain value, its first and second derivatives w.r.t the displacements and the internal variable `alpha_`.
+- Add Q1Hn and Q1HTn elements ([#363](https://github.com/ikarus-project/ikarus/pull/363))
+- Add wrappers for solving the generalized eigenvalue problem with sparse and dense matrices with Eigen and Spectra ([#368](https://github.com/ikarus-project/ikarus/pull/368)).
 
 ## Release v0.4 (Ganymede)